backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
dj-database-url==2.1.0
python-decouple==3.8
whitenoise==6.5.0
gunicorn==21.2.0